iska_python_project/pyproject.toml
[project]
name = "iska-python-project"
version = "0.2.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity-broker>=1.3.0",
    "fabric-cicd>=0.1.25",
    "ms-fabric-cli>=1.0.1",
    "ruff>=0.12.11",
]
[tool.setuptools]
packages = ["iska_python_project"]
